+++ src/testing-utils/if-not-type/if-not/if-not.ts
@@ -3,6 +3,9 @@
  */
 
 /**
+ * @template Condition The boolean condition against which the If statement checks.
+ * @template IfTrue The resulting type if `Condition` extends `false`. Default: `true`.
+ * @template IfFalse The resulting type if `Condition` extends `true`. Default: `false`.
  * @description IfNot type factory is used to create other, more specific IfNotType types.
  * @example Example.
  */