+++ src/testing-utils/is-type/booleans/is-boolean-literal.ts
@@ -1,11 +1,16 @@
 /**
- * @file Returns true if a boolean literal (`true` or `false`) is passed to it.
+ * @file Returns true if a boolean literal (`true` OR `false`) is passed to it.
  * Returns false if `boolean`, `true | false`, or anything else is passed to it.
  */
 
-import type { IsAny } from '../any-never-unknown/is-any';
-import type { IsNever } from '../any-never-unknown/is-never';
-import type { IsUnknown } from '../any-never-unknown/is-unknown';
+/* eslint-disable @typescript-eslint/no-unused-vars -- Need this for JSDoc @link tags. */
+import type { IsBoolean } from './is-boolean';
+import type { IsBooleanBroad } from './is-boolean-broad';
+import type { IsTrue } from './is-true';
+import type { IsFalse } from './is-false';
+/* eslint-enable @typescript-eslint/no-unused-vars -- Types actually used for IsBooleanLiteral should come after this line. */
+
+import type { IsAnyOrNeverOrUnknown } from '../any-never-unknown/is-any-or-never-or-unknown';
 
 /**
  * @template TestedType
@@ -15,6 +20,12 @@ import type { IsUnknown } from '../any-never-unknown/is-unknown';
  * - `false`;
  *
  * Returns `false` otherwise.
+ *
+ * Broader variation: {@link IsBoolean}.
+ *
+ * Narrower variations: {@link IsTrue}, {@link IsFalse}.
+ *
+ * Counterpart: {@link IsBooleanBroad}.
  * @example
  * ```
  * // Only boolean literals (true OR false) will return true:
@@ -32,17 +43,20 @@ import type { IsUnknown } from '../any-never-unknown/is-unknown';
  * type IsBooleanLiteral_String = IsBooleanLiteral<string>; // false
  * type IsBooleanLiteral_Number = IsBooleanLiteral<number>; // false
  * ```
+ * @see
+ * Types used under the hood:
+ * - IsType: {@link IsAnyOrNeverOrUnknown}.
  */
-export type IsBooleanLiteral<TestedType> = IsAny<TestedType> extends true
-  ? false
-  : IsNever<TestedType> extends true
-  ? false
-  : IsUnknown<TestedType> extends true
-  ? false
-  : boolean extends TestedType
-  ? false
-  : TestedType extends true
-  ? true
-  : TestedType extends false
-  ? true
-  : false;
+export type IsBooleanLiteral<TestedType> =
+  // Necessary to test for `any`, `never`, and `unknown` early on.
+  IsAnyOrNeverOrUnknown<TestedType> extends true
+    ? false
+    : // The condition below tests for broad boolean type,
+    // and hence needs to be excluded.
+    boolean extends TestedType
+    ? false
+    : TestedType extends true
+    ? true
+    : TestedType extends false
+    ? true
+    : false;